Added revealTickets method
Added the new revealTickets method and the accompanying error (426). Also includes tests for the method (revealing a single ticket and multiple tickets) and the error.

Updated the RandomOrgRANDOMORGError class to include the error code as a (private) property, accessible by using the getCode() method.

Also updated the links in the descriptions.